Add universal loading state to Header component, user server component on Service initial view to test
