+++ src/app/home/home.scss
@@ -88,23 +88,6 @@
   overflow: hidden;
 }
 
-/* Title and avatar w/notifications */
-.Header {
-  background: white;
-  overflow: hidden;
-  
-  height: 54px;
-  width: 100%;
-
-  padding: 0 25px;
-
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  border-bottom: 1px solid #DDDDDD;
-}
-
 .SecondaryHeader { 
  width: 100%;
  height: 58px;
