+++ src/app/home/services/loading.tsx
@@ -1,0 +1,9 @@
+import { Header } from "@/components/Header"
+
+export default function Loading() {
+  
+  return <div>
+    <Header text="Services" loading />
+    <p>loading...</p>
+  </div>
+}
